Add remote arklone parent directory to rclone scripts

+++ tests/rclone/scripts/send-arkos-backup.sh
@@ -3,17 +3,16 @@
 # by ridgek
 # Released under GNU GPLv3 license, see LICENSE.md.
 
+# Only run if on ArkOS
+if ! grep "title=" "/usr/share/plymouth/themes/text.plymouth" | grep "ArkOS"; then
+	exit
+fi
+
 source "/opt/arklone/config.sh"
 
 ###########
 # MOCK DATA
 ###########
-# Create some test directories and files
-ARKLONE[backupDir]="/dev/shm/backup"
-mkdir "${ARKLONE[backupDir]}"
-
-touch "${ARKLONE[backupDir]}/arkosbackup.tar.gz"
-
 # Mock test rclone.conf
 ARKLONE[rcloneConf]="/dev/shm/rclone.conf"
 
@@ -23,36 +22,38 @@ type = local
 nounc = true
 EOF
 
-# @todo where does rclone write to when remote is local
-# 	and remote is passed like remote:someDir/ ?
 ######
 ## RUN
 ######
-## Run script
-#. "${ARKLONE[installDir]}/rclone/send-arkos-backup.sh"
-#
-#[ $? = 0 ] || exit 70
-#
+# Run rclone in /dev/shm because remote is local filesystem
+cd "/dev/shm"
+
+# Run script
+. "${ARKLONE[installDir]}/rclone/send-arkos-backup.sh"
+
+[ $? = 0 ] || exit 70
+
 #########
 ## TEST 1
 #########
-## Log file exists
-#[ -f "${ARKLONE[backupDir]}/arkosbackup.log" ] || exit 72
-#
-# echo "TEST 1 passed."
-#
+# Log file exists
+[ -f "/roms/backup/arkosbackup.log" ] || exit 72
+
+ echo "TEST 1 passed."
+
 #########
 ## TEST 2
 #########
-## Backup file was synced
-#[ -d "?" ] || exit 72
-#[ -f "?/arkosbackup.tar.gz" ] || exit 72
-#
-# echo "TEST 2 passed."
-#
+# Backup file was synced
+[ -f "/dev/shm/arklone/ArkOS/arkosbackup.tar.gz" ] || exit 72
+
+ echo "TEST 2 passed."
+
 ###########
 ## TEARDOWN
 ###########
-rm -rf "${ARKLONE[backupDir]}"
 rm "${ARKLONE[rcloneConf]}"
+rm "/roms/backup/arkosbackup.log"
+rm "/roms/backup/arkosbackup.tar.gz"
+rm -rf "/dev/shm/arklone"
